Supermatter Fusion Stuff
Oh hell, there's a bunch of stuff...

- Supermatter efficiency reduced from 1.0 to 0.4.

- Smacking supermatter shards or entire supermatters will raise the
smlevel

- Supermatter explosion and delamination nerfed. Explosion scales with
smlevel

- Supermatters of different stages now glow different colors.

- Fixes tongs being able to pick up shards when they are already holding
something.

- Hitting a shard with tongs carrying a shard, will cause the shards to
fuse.

- High-Level SM growths have a chance to delaminate into lower-level
shards.

- SM is able to take more damage when at higher levels. This makes it
delaminate faster, and seem more unstable.

- For each stage of the SM, the explosion point is raised by 250.

- Supermatters with a level above 1 have a chance for "critical
failure". The damage inflicted by, and probability of a failure
increases with smlevel.

- Supermatters are drawn to eachother, and try to pull shards into
themselves.

- The amount of Nitrogen needed to stabilize the oxygen reaction
increases for each stage. At stage 8, the supermatter is just barely
able to react with the environments air (79% nitrogen; 21% oxygen),
above stage 8, the supermatter will react with station air.

- Supermatter loses power slower at higher levels. This makes it
possible to produce more power, but also makes it harder to maintain.

- The supermatter produces more phoron per mole of oxygen reacted at
higher stages.

- The supermatters "gas efficiency" rises with its level, capping out at
stage 10. At stage 10, it is able to react with all the oxygen in the
environment per-process, and produces 5 moles of phoron for every 1 mole
of oxygen.

- radiation and hallucination is buffed with stronger SM's

- The SM is able to power collectors 2 tiles further for each level.

- Delaminations now use smlevel. A Delamination of lvl 2 is able to
produce shards of level 2 and under.