Lightswitch move
The lightswitch at 126,146,1 appears to not have been moved when the
bar was updated and looks like it's in the toilet next door. I've moved
it one tile left so it's now in the entrance to the toilet.

Another great example of when one minor change in the byond editor
makes a ton of changes to the dmm file.