Added sector flags
Sectors now are attributed flags, which determine what kind of sector
they are. Currently there's 5 different types of flags

SECTOR_KNOWN: Is this sector known? Basically makes a sector visible on
the overmap
SECTOR_STATION: Is this sector a part of the station?
SECTOR_ALERT: Is this sector affected by alert levels?
SECTOR_LOCAL: Is this sector even present on the overmap?
SECTOR_ADMIN: Is this sector admin only?