SMES Access changes
A number of SMES rooms across the station seem to have different access
or allow that department into the SMES room, the most concerning of
which is Civilian east which allows (you guessed it) civilians in to
make changes.

The only rooms that seem to be correct were civilian west (near
arrivals) and Security, I've now changed all the doors to be
engineering doors and changed the access to 11 (Engine Setup) and 24
(Atmospherics). The engineering SMES room seemed to have 10 and 24 so
I've changed that to now match the others.

I've not changed the bar's SMES as it's on a different Z level and I
think there's more changes to that area i'd like to make as well such
as moving the lower bar so just the barman has access to a small
maintenance tunnel linking the upper and lower bar.