Adds Chemical Spills
Adds base for chemicals spills. This increases the persistence of
chemicals in the game. Instead of being able to spill a chemical onto
the ground in order to delete it, it will leave a chemical spill, which
will react with other chemicals dumped on it, or flood over onto other
tiles if 'filled' enough.

Ideally this will include and replace dirt, tracks, blood, oil, radium,
foam, paint, etc. And will allow for some interesting scenarios and
combinations.

Such as drying agent + water can be placed in the mop bucket instead of
just plain-ol'-water to speed up the rate at which the floor dries.

Or assuming paint becomes a reagent, to dry paint more quickly. Or that
different 'pigments' can be added to the paint in the form of chemicals
in order to change the colour of it.

This adds a base for flooding mechanics, so that a room can be flooded
with water, or any other chemical. As well as a chemical process/ticker,
for better handling metabolism, water drying, blood, decomposition,
slow-acting acids, etc.

Also adds alpha for chems. Dunno why we still didnt have it yet before.