Merge dev into master (#1093)
* Evidence bag TG port
Ports evidence bag code from /tg/, had to add some necessary procs to some files.
Evidence bags are now afterattack() and attackby() based instead of mousedrop().

* Confirmation for if you want to put an item in a bag or not. Should prevent any unwanted bagging.

* Fix for walking away and teleporting objects into the bag

* HoP QoL changes

* Buildable Display Cases
This adds display cases buildable using wood (for the chassis) and glass (added to the chassis) and airlock electronics. The airlock electronics access is the access required to open and close the display case. Just use an ID on it to toggle.
The item in the display case IS visible.

* Randomized Plant Harvests
This makes botany plants give random harvests each time instead of a static harvest. This will stop the smartfridge from being loaded with oranges and apples that have a meager amount of recipes limited to pies and cakes and the very same smartfridge having almost no potatoes or tomatoes which are used in pretty much most of the recipes.
Also I made harvesting a plant give you a message telling you how much you harvested. Also gives you a fluff message if the harvest was at the minimum or maximum.

* Merge conflict fix

* why is this still here?

* Put propaganda contract items into backpack if possible

* Poison contract

Poison contract which random antags can take as well
Also adds the reagent dicardine

* Fix poison contract compile error

* Engine project (#1070)

Tokamak/Fusion engine.
This contains:
Fusion engine, controller for the fusion engine, fusion ball and all the controller compo

* Added support for field generators

* tokamak engine implementatin

* sprite change

* Added button support for arc emitters

* Resolved some spelling errors in the Tokamak engine components

* More name changeing

* More name updating

* addition of Arc Gun (lighting gun) Fires bolts of lightning (very un balanced). Also adds it to the dispensers on the thunderdome.

* Updating license document with mention of licence of used libraries.

* Readme licence update 2

* Fixed bug with components not beeing ready when they should be. Arc emitter damage to heat balanced.

* More balances, map update regarding vacume in engine room.

* Map, engine airlock update. IDD and Tokamak controll console tag update.

* medical revamp

* engine tweaks

* Remaking the interaction menu to nanoui. Removing all dev from tokamak.