This change primarily changes the bridge, adding more consoles to the control center, adding an authenticator to the conference room, as well as minor changes to the Captain's Office. There is also a cautery and syringe box placed into the robotic lab and chemistry lab respectively.
The command center changes allow for it to act as a more proper and authoritative command center, with the authenticator allowing a Head with a damaged, unpowered, or overtaken office to still swipe, and the Captain's Office changes merely being aesthetic. The items placed into the labs are things that should've been there in the first place.